Everything in doubles... but it does still not work!
5 bucks for the one who finds the error...

+++ App/RefinerServer/RefinerServer/ImageRepresentation.h
@@ -15,8 +15,8 @@
 	{
 	public:
 		ImageRepresentation(std::string filename,
-			Eigen::Matrix4f pCameraViewTransform,
-			Eigen::Matrix4f pCameraProjectionTransform);
+			Eigen::Matrix4d pCameraViewTransform,
+			Eigen::Matrix4d pCameraProjectionTransform);
 
 		/// <summary>
 		/// compute 3D projection of 2D point "pixel" onto surface defined by normal "surface_normal" and vertex "vertex"
@@ -26,19 +26,19 @@
 		/// <summary>
 		/// Compute corrsponding patch in camera 2 given the patch in camera 1 (via reprojection and projective unwarping)
 		/// </summary>
-		float computeDistortedPatchCorrelation(ImageRepresentation& image2, Eigen::Vector3d surface_normal, Eigen::Vector3d vertex, cv::Size patch_size, int colorFlag);
+		double computeDistortedPatchCorrelation(ImageRepresentation& image2, Eigen::Vector3d surface_normal, Eigen::Vector3d vertex, cv::Size patch_size, int colorFlag);
 		
 
 		/// <summary>
 		/// Compute corrsponding patch in camera 2 given the patch in camera 1 (without unwarping)
 		/// </summary>
-		float computePatchCorrelation(ImageRepresentation& image2, Eigen::Vector3d surface_normal, Eigen::Vector3d vertex, cv::Size patch_size);
+		double computePatchCorrelation(ImageRepresentation& image2, Eigen::Vector3d surface_normal, Eigen::Vector3d vertex, cv::Size patch_size);
 
 
 		/// <summary>
 		/// Compute view quality
 		/// </summary>
-		float getViewQuality(Eigen::Vector3d vertex, Eigen::Vector3d normal, ImageRepresentation& image2);
+		double getViewQuality(Eigen::Vector3d vertex, Eigen::Vector3d normal, ImageRepresentation& image2);
 
 
 		const std::string filename;
@@ -46,6 +46,6 @@
 		cv::Mat ocvImage;
 		unsigned width, height;
 		int x_size, y_size;
-		Eigen::Matrix4f CameraViewTransform;
-		Eigen::Matrix4f CameraProjectionTransform;
+		Eigen::Matrix4d CameraViewTransform;
+		Eigen::Matrix4d CameraProjectionTransform;
 	};
